Use size_t for Repeats in AddTimer

+++ Client/mods/deathmatch/logic/lua/CLuaTimerManager.cpp
@@ -131,7 +131,7 @@ CLuaTimer* CLuaTimerManager::GetTimerFromScriptID(uint uiScriptID) const
     return IsValidTimer(pLuaTimer) ? pLuaTimer : nullptr;
 }
 
-CLuaTimer* CLuaTimerManager::AddTimer(const CLuaFunctionRef& iLuaFunction, CTickCount llTimeDelay, unsigned int uiRepeats, CLuaArguments Arguments)
+CLuaTimer* CLuaTimerManager::AddTimer(const CLuaFunctionRef& iLuaFunction, CTickCount llTimeDelay, size_t uiRepeats, CLuaArguments Arguments)
 {
     // Check for the minimum interval
 #if LUA_TIMER_MIN_INTERVAL
